Extract `PrimitiveValue` to a separate file
