+++ src/main/kotlin/sigma/PrimitiveValue.kt
@@ -1,0 +1,12 @@
+package sigma
+
+sealed class PrimitiveValue : Value() {
+    final override fun equals(other: Any?): Boolean {
+        if (other !is Value) return false
+        return isSame(other)
+    }
+
+    final override fun hashCode(): Int = 0
+
+    abstract fun isSame(other: Value): Boolean
+}
